// src/app/ui/categories/productActions.enum.ts
export enum PRODUCT_ACTIONS {
    ADD_PRODUCT = 'ADD_PRODUCT',
    REMOVE_PRODUCT = 'REMOVE_PRODUCT',
    QUANTITY_PLUS = 'QUANTITY_PLUS',
    QUANTITY_MINUS = 'QUANTITY_MINUS',
    BUY = 'BUY',
    EDIT = 'PRODUCT_EDIT',
    GET_FORM_FIREBASE = 'GET_FROM_FIREBASE_PRODUCT',
    UNBUY_ALL_PRODUCTS = 'UNBUY_ALL_PRODUCTS'
}